# api/src/main/resources/swagger.yml
openapi: 3.0.1
info:
  version: ''
  title: Link API
servers:
  - url: ''
    description: This server
security:
  - oauth: []
  - basic: [ ]
...
paths:
...
  /api/{tenantId}/report/$generate:
    post:
      tags:
        - Reports
      summary: Generate a report for given criteria
      description: Generates a report based on the specified measure and reporting period criteria. Report generation can take time to respond, as it involves querying the configured EHR for data related to the patients identified for the criteria.
      operationId: generateReport_1
      parameters:
        - name: tenantId
          in: path
          description: The id of the tenant
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateRequest'
      responses:
        '200':
          description: successful response
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Report'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/{tenantId}/report/$generateMultiMeasure:
    post:
      tags:
        - Reports
      summary: Generate a report for given criteria
      description: Generates a report based on the specified measure and reporting period criteria. Report generation can take time to respond, as it involves querying the configured EHR for data related to the patients identified for the criteria.
      operationId: generateReport_2
      parameters:
        - name: tenantId
          in: path
          description: The id of the tenant
          required: true
          schema:
            type: string
        - name: multiMeasureBundleId
          in: query
          description: The id of the multi measure package/bundle
          required: true
          schema:
            type: string
        - name: periodStart
          in: query
          description: The start date of the reporting period, formatted as a FHIR datetime
          required: true
          schema:
            type: string
        - name: periodEnd
          in: query
          description: The end date of the reporting period, formatted as a FHIR datetime
          required: true
          schema:
            type: string
        - name: regenerate
          in: query
          description: Whether to regenerate the report if a report with the same criteria already exists
          required: true
          schema:
            type: boolean
      responses:
        '200':
          description: successful response
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Report'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/{tenantId}/report/{reportId}/$send:
    post:
      tags:
        - Reports
      summary: Send/submit the specified report to the consumer/recipient
      operationId: send
      parameters:
        - name: tenantId
          in: path
          description: The id of the tenant
          required: true
          schema:
            type: string
        - name: reportId
          in: path
          description: The logical id of the report to submit
          required: true
          schema:
            type: string
        - name: download
          in: query
          description: Whether to echo the report to the client
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: successful response
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/{tenantId}/report/{reportId}/aggregate:
    get:
      tags:
        - Reports
      summary: Get the aggregate measure report(s) generated the specified report id
      operationId: getReportAggregates
      parameters:
        - name: tenantId
          in: path
          description: The id of the tenant
          required: true
          schema:
            type: string
        - name: reportId
          in: path
          description: The id of the report
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A list of FHIR MeasureReports
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FHIR.MeasureReport'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/{tenantId}/report/{reportId}/patientList:
    get:
      tags:
        - Reports
      summary: Get the patient list(s) used to generate the specified report id
      operationId: getReportPatientLists
      parameters:
        - name: tenantId
          in: path
          description: The id of the tenant
          required: true
          schema:
            type: string
        - name: reportId
          in: path
          description: The id of the report
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful response
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PatientList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/{tenantId}/report/{reportId}/individual:
    get:
      tags:
        - Reports
      summary: Get all individuals' patient measure reports generated for the specified report id
      operationId: getPatientMeasureReports
      parameters:
        - name: tenantId
          in: path
          description: The id of the tenant
          required: true
          schema:
            type: string
        - name: reportId
          in: path
          description: The id of the report
          required: true
          schema:
            type: string
      responses:
        '200':
          description: All individuals' FHIR MeasureReport
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/FHIR.MeasureReport'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/{tenantId}/report/{reportId}/individual/{patientMeasureReportId}:
    get:
      tags:
        - Reports
      summary: Get an individual's patient measure report generated for the specified report id
      operationId: getPatientMeasureReport
      parameters:
        - name: tenantId
          in: path
          description: The id of the tenant
          required: true
          schema:
            type: string
        - name: reportId
          in: path
          description: The id of the report
          required: true
          schema:
            type: string
        - name: patientMeasureReportId
          in: path
          description: The id of the patient measure report
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A individual's FHIR MeasureReport
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/FHIR.MeasureReport'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
...
    PatientId:
      type: object
      properties:
        reference:
          type: string
        identifier:
          type: string
    PatientList:
      type: object
      properties:
        id:
          type: string
        periodStart:
          type: string
        periodEnd:
          type: string
        measureId:
          type: string
        lastUpdated:
          type: string
        patients:
          type: array
          items:
            $ref: '#/components/schemas/PatientId'
    GenerateRequest:
      type: object
      properties:
        regenerate:
          type: boolean
        bundleIds:
          type: array
          items:
            type: string
        periodStart:
          type: string
        periodEnd:
          type: string
    Report:
      type: object
      properties:
        id:
          type: string
        measureId:
          type: array
          items:
            type: string
        periodStart:
          type: string
        periodEnd:
          type: string
        status:
          type: string
          enum:
            - Draft
            - Submitted
        submittedTime:
          type: string
        generatedTime:
          type: string
        version:
          type: string
        patientLists:
          type: array
          items:
            type: string
        aggregates:
          type: array
          items:
            type: string
...
    FHIR.MeasureReport:
      type: object
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
  securitySchemes:
    oauth:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: ''
          tokenUrl: ''
          scopes: { }
    basic:
      type: http
      scheme: basic
  responses:
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'